+++ Runtime/CloudAuthProvider.cs
@@ -1,0 +1,10 @@
+namespace Wagenheimer.CloudSave
+{
+    public enum CloudAuthProvider
+    {
+        Anonymous,
+        GooglePlayGames,
+        Apple,
+        AppleGameCenter
+    }
+}